requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.13.0
async-timeout==5.0.1
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
contourpy==1.3.2
cuda-bindings==13.2.0
cuda-pathfinder==1.5.4
cuda-toolkit==13.0.2
cycler==0.12.1
datasets==4.8.5
dill==0.4.1
exceptiongroup==1.3.1
filelock==3.29.0
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.13.0
idna==3.13
Jinja2==3.1.6
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.9
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
networkx==3.4.2
numpy==2.2.6
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
opencv-python==4.13.0.92
packaging==26.2
pandas==2.3.3
pillow==12.2.0
propcache==0.4.1
pyarrow==24.0.0
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
rich==15.0.0
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
triton==3.6.0
typer==0.25.1
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
xxhash==3.7.0
yarl==1.23.0
